refactor(daemon): re-org the loaders of configs
