+++ crates/core/Cargo.toml
@@ -19,6 +19,7 @@ futures-util.workspace = true
 schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+serde_yaml.workspace = true
 smallvec.workspace = true
 tokio.workspace = true
 tracing.workspace = true
