+++ crates/core/src/agent/parser.rs
@@ -1,9 +1,10 @@
-//! Agent markdown loading.
+//! Agent markdown parsing.
+//!
+//! Parses YAML frontmatter + Markdown body into an [`AgentConfig`].
 
-use crate::loader::split_yaml_frontmatter;
+use crate::agent::config::AgentConfig;
+use crate::utils::split_yaml_frontmatter;
 use serde::Deserialize;
-use std::path::Path;
-use wcore::AgentConfig;
 
 /// YAML frontmatter for agent markdown files.
 #[derive(Deserialize)]
@@ -38,30 +39,4 @@ pub fn parse_agent_md(content: &str) -> anyhow::Result<AgentConfig> {
     };
 
     Ok(config)
-}
-
-/// Load all agent markdown files from a directory.
-///
-/// Each `.md` file is parsed with [`parse_agent_md`]. Non-`.md` files are
-/// silently skipped. Entries are sorted by filename for deterministic ordering.
-/// Returns an empty vec if the directory does not exist.
-pub fn load_agents_dir(path: &Path) -> anyhow::Result<Vec<AgentConfig>> {
-    if !path.exists() {
-        tracing::warn!("agent directory does not exist: {}", path.display());
-        return Ok(Vec::new());
-    }
-
-    let mut entries: Vec<_> = std::fs::read_dir(path)?
-        .filter_map(|e| e.ok())
-        .filter(|e| e.path().extension().is_some_and(|ext| ext == "md"))
-        .collect();
-    entries.sort_by_key(|e| e.file_name());
-
-    let mut agents = Vec::with_capacity(entries.len());
-    for entry in entries {
-        let content = std::fs::read_to_string(entry.path())?;
-        agents.push(parse_agent_md(&content)?);
-    }
-
-    Ok(agents)
 }
