+++ hook/cron/Cargo.toml
@@ -17,6 +17,7 @@ wcore.workspace = true
 compact_str.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+serde_yaml.workspace = true
 cron.workspace = true
 chrono.workspace = true
 tokio.workspace = true
